Rename pictogram comps to avoid name clashes
