+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nethserver/ns8-ui-lib",
-  "version": "0.0.48",
+  "version": "0.0.49",
   "description": "Vue.js library for NethServer 8 UI",
   "keywords": [
     "nethserver",
