+++ src/lib-components/pictograms/BulldozerPictogram.vue
