+++ src/lib-components/pictograms/ExclamationMarkPictogram.vue
