+++ src/lib-components/pictograms/GearPictogram.vue
