+++ src/lib-components/pictograms/GroupPictogram.vue
