+++ src/lib-components/pictograms/HardDrivePictogram.vue
