+++ src/lib-components/pictograms/LovePictogram.vue
