+++ src/lib-components/pictograms/WarningPictogram.vue
